fix: Pop all events from stack
It can happen that the dialog is not able to render before the next event is received. Before that patch that could have lead to some random dangling events that get shown once the root screen gets rebuild.

This patch puts new events in a last in last out order and on rendering the task dialog pops all events from the queue. This way we prevent having dangling events.

fixes https://github.com/get10101/10101/issues/2506