chore: Add postgrest-coordinator container to docker-compose.yml
This server can be used during local development to query the
coordinator database conveniently e.g.

```
$ curl http://localhost:3002/dlc_protocols | jq .
[
  {
    "id": 1,
    "protocol_id": "fb4c6f27-db9d-4b2d-8539-6e609e1559b9",
    "previous_protocol_id": null,
    "channel_id": "f689230c5477176961301cb22093cfd52050a52356812039f9c298fe8859eba2",
    "contract_id": "972c5fcafbd6bbbf874a9a5cd44d9ddd05a2e3f3ae96d5bd771e5af1b8cae1a6",
    "protocol_state": "Success",
    "trader_pubkey": "02e9f7a0e0d6eeab989dcbb4e91f0bf95f583b09d8d324cc9635807e844eea82bc",
    "timestamp": "2024-05-13T05:32:36.452958+00:00",
    "protocol_type": "open-channel"
  }
]
```

It can also be used in the e2e tests to easily assert on the state of
the coordinator, without needing to add dedicated HTTP API endpoints.
It might even be used to modify the coordinator state to create
interesting test scenarios.