feat: Let the coordinator open single-funded channels
With the `TradeAction::OpenSingleFundedChannel`, the coordinator can
choose to open a DLC channel with a trader, without using any trader
UTXOs. This is useful if the coordinator wants to onboard trading
through alternate means, such as Lightning.

`rust-dlc` now allows to configure the TX fee split between the offer
party and the accept party.

The feature has been tested locally, but is currently disabled as we
need a different flow to trigger the creation of a single-funded DLC
channel.

Also, the DLC channel is now 0-conf: the trader can keep trading as
soon as the fund transaction is found in mempool.