feat: fund channel flow
We introduce a new flow to fund the first position and with it the first channel. The idea is that the user does not have to have any on-chain funds before opening the first position.
The flow is divided into two parts:
1) open a position using on-chain funds:
if the user has already on-chain funds, then he can opt to use said funds and open the channel immediately
2) open a position using external funds:
the user has the option to fund the wallet using an external wallet such as a lightning wallet, another on-chain wallet, or maybe even using a hardware wallet in the future.
In this patch, we only introduce funding using an on-chain transaction. The app waits until it sees the funds incoming and then proceeds with the channel creation.