fix(app-test): workaround for fake `GoRouter` which cant navigate
We can't navigate using the mocked router (I don't understand why). Which means, we can't navigate to the new channel configuration flow.
Because of this, I've split up the tests into more fine granular tests. We test now each widget separately.
This has the advantage that our tests are smaller and easier to understand.
It seems like that if we want to test navigation, we should write integration tests instead.