fix: Ignore missing order id when rejecting a rollover offer
In case of a rollover we do not have an order id or and order in filling. As a result this logic failed when trying to reject a rollover offer.

The solution is still hacky as with the correct context information we shouldn't even try to set the order to failed, but that context is missing when we are reconnecting in the `on_connect` function and finding a dlc channel in state `RenewOffered`.