fix(coordinator): Remove insane compile_error!
I was trying to fail at compile time based on a runtime value (the
settings). This is dumb.

We now fail at runtime if we try to use a fixed index price in release
mode. Production always runs on release, so this exists to prevent
using a fixed index price in production.

Obviously, release does not always equal production, but using the
real index price from BitMEX is fine in most scenarios (except for
the e2e tests and some instances of local testing).