fix: Drop foreign key constraint on order id
The order and the matches are persisted by async actors. It can happen that the market order has not yet been persisted when the match is processed.

Thus it could happen that the matches table is persisted before the order has been persisted, which resulted in a foreign key violation. By dropping the foreign key we don't care if the order is already persisted and can continue with eventual consistency.

I think this is the better option than to synchronize the two inserts as we are anyways aiming for decoupling the orderbook from the coordinator.