feat: Sync channel eagerly when proposing a rollover
Analogue to what we do when opening a position. We check if the contract is confirmed. If not, we sync eagerly - thus increasing the chances that the rollover proposal will succeed.

This covers a small edge case where the user opened a position right before the rollover window and the coordinator hasn't synced yet.