fix: Only accept a rollover offer if the contract has been confirmed.
Syncs the wallet eagerly if the channel is not yet confirmed.

If the contract is not confirmed, the rollover offer gets rejected. This fixes an issue where the app otherwise risk ending up in an intermediate state as it can't accept the `RenewConfirm` message.

see https://github.com/get10101/10101/issues/2630